Fix dataset skeleton. Banner styles and images. Show Network diagram expand button
